(bug fix)获取核数返回值不能为int，应该为int64，否则transfer会过滤掉

+++ modules/agent/funcs/cpustat.go
@@ -163,8 +163,8 @@ func CpuPrepared() bool {
 	defer psLock.RUnlock()
 	return procStatHistory[1] != nil
 }
-func CpuCores() int {
-	return runtime.NumCPU()
+func CpuCores() uint64 {
+	return uint64(runtime.NumCPU())
 }
 
 func CpuMetrics() []*model.MetricValue {
